Flattened `scriptdata` mod
